moved helper to separate file
